Add files via upload
In this project I will be assigning input function to 4 gpio pins in nodemcu. 3 of them are used for giving input in morse and 4th pin is used to print the array till now. Nodemcu doesn't supports keyboard libraries and perform like 32u4 s , so this code will help me to do that without those dependencies.
Also these buttons are edge triggered.
I have some of the problems int the code because it is not performing the tasks in which it is supposed to, I get different output.
Need you help as soon as possible, please.
Circuit is very simple, the gpio pins only monitor the circuit completion.
I also have included more description inside the code,